[Docs] First pass attempt at getting pdf and epub output built.
I spent some time today attempting to coerce Sphinx to produce a pdf.
Getting pdf2latex functioning on Windows is not as trivial as you might
hope for!

End result: use https://miktex.org/download, some futzing of package
manager required to install required packages.

Finally managed to produce a pdf
I had to change references to some names as we had originally
presented them in a unicode font which pdf2latex was dying on.
There was a weird quote issue in one file that required tweaking.

I'm not happy with the output for pdf or epub as it stands: a lot of our
monospace code blocks are too wide for the output format and are just
truncated.

See GitHub issue #1996